Refactor module locations
