+++ Causal_Web/engine/logging/log_interpreter.py
@@ -2,7 +2,7 @@ import json
 import os
 from typing import Dict, List
 
-from ..causal_analyst import CausalAnalyst
+from .causal_analyst import CausalAnalyst
 from ..models.base import OutputDirMixin, JsonLinesMixin
 from ...config import Config
