+++ Causal_Web/engine/models/meta_node.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Dict, List, Optional
 
-from .models.node import Node
+from .node import Node
 
 
 @dataclass
